Rename config.py into common.py
