+++ python/paddle/v2/dataset/common.py
